feat: adjust program

+++ graph/Graph.cpp
@@ -29,20 +29,20 @@ Graph::addEdge(VID _from, VID _to, LABEL _elb)
 void
 Graph::buildCSR()
 {
-	int verSize = this->vSize();
-	labels = new int[verSize];
-	for(int i = 0; i < verSize; i ++)
+	long verSize = this->vSize();
+	labels = new long[verSize];
+	for(long i = 0; i < verSize; i ++)
 		labels[i] = vertices[i].label;
 
-	inRowOffset = new int[verSize + 1];
+	inRowOffset = new long[verSize + 1];
 	inRowOffset[0] = 0;
-	outRowOffset = new int[verSize + 1];
+	outRowOffset = new long[verSize + 1];
 	outRowOffset[0] = 0;
 
-	int inTotalNum = 0;
-	int outTotalNum = 0;
+	long inTotalNum = 0;
+	long outTotalNum = 0;
 
-	for(int i = 0; i < verSize; i ++)
+	for(long i = 0; i < verSize; i ++)
 	{
 		inTotalNum += this->vertices[i].in.size();
 		outTotalNum += this->vertices[i].out.size();
@@ -52,24 +52,24 @@ Graph::buildCSR()
 		sort(this->vertices[i].out.begin(),this->vertices[i].out.end(),cmpNeighbor);
 	}
 
-	inColOffset = new int [inTotalNum];
-	inValues = new int [inTotalNum];
+	inColOffset = new long [inTotalNum];
+	inValues = new long [inTotalNum];
 
-	outColOffset = new int[outTotalNum];
-	outValues = new int [outTotalNum];
+	outColOffset = new long[outTotalNum];
+	outValues = new long [outTotalNum];
 
-	int inPos = 0;
-	int outPos = 0;
-	for(int i = 0; i < this->vSize(); i++)
+	long inPos = 0;
+	long outPos = 0;
+	for(long i = 0; i < this->vSize(); i++)
 	{
 	//	printf("the %d th node's in size is %d,out size is %d\n",i,this->vertices[i].in.size(),this->vertices[i].out.size());
-		for(int j = 0; j < this->vertices[i].in.size(); j ++)
+		for(long j = 0; j < this->vertices[i].in.size(); j ++)
 		{
 			inColOffset[inPos] = ((this->vertices[i].in)[j]).vid;
 			inValues[inPos] = ((this->vertices[i].in)[j]).elb;
 			inPos ++;
 		}
-		for(int j = 0; j < this->vertices[i].out.size(); j ++)
+		for(long j = 0; j < this->vertices[i].out.size(); j ++)
 		{
 			outColOffset[outPos] = ((this->vertices[i].out)[j]).vid;
 			outValues[outPos] = ((this->vertices[i].out)[j]).elb;
@@ -95,17 +95,17 @@ Graph::buildCSR()
 void 
 Graph::buildReverse()
 {
-	reverseLabel = new int [nodeLabelNum + 1];
+	reverseLabel = new long [nodeLabelNum + 1];
 	reverseLabel[0] = 0;
 //	printf("diffLabelNum is %d\n",diffLabelNum[0]);
-	for(int i = 1; i <= nodeLabelNum; i++)
+	for(long i = 1; i <= nodeLabelNum; i++)
 		reverseLabel[i] = reverseLabel[i-1] + diffLabelNum[i-1];
-	reverseValues = new int [reverseLabel[nodeLabelNum]];
-	int curLabelPos[nodeLabelNum];
-	memcpy(curLabelPos,reverseLabel,sizeof(int)*nodeLabelNum);
-	for(int i = 0; i < vertices.size(); i ++)
+	reverseValues = new long [reverseLabel[nodeLabelNum]];
+	long curLabelPos[nodeLabelNum];
+	memcpy(curLabelPos,reverseLabel,sizeof(long)*nodeLabelNum);
+	for(long i = 0; i < vertices.size(); i ++)
 	{
-		int label = vertices[i].label - 1;
+		long label = vertices[i].label - 1;
 		reverseValues[curLabelPos[label]] = i;
 		curLabelPos[label] ++;
 	}
